+++ infrastructure/quick-deploy/aws/armonik/variables.tf
@@ -86,6 +86,8 @@ variable "control_plane" {
     image_pull_secrets = string
     node_selector      = any
     annotations        = any
+    # KEDA scaler
+    hpa                = any
   })
 }