+++ CHANGELOG.md
@@ -48,6 +48,7 @@ Added
 * Unified API delivery with User, Admin and monitoring API
 * Set fs.inotify.max_user_instances to 8192 in worker nodes on AWS
 * Expose the parameters of the cluster autscaler's Helm chart in Terraform sources
+* Add KEDA HPA for the ArmoniK control plane
 
 Changed
 -